TestUDM/GroupCommandTest.cs
using UDM;
using Microsoft.VisualStudio.TestTools.UnitTesting;
using System;
using System.Collections.Generic;

namespace TestUDM
{
    
    
    /// <summary>
    ///This is a test class for GroupCommandTest and is intended
    ///to contain all GroupCommandTest Unit Tests
    ///</summary>
    [TestClass()]
    public class GroupCommandTest
    {


        private TestContext testContextInstance;

        /// <summary>
        ///Gets or sets the test context which provides
        ///information about and functionality for the current test run.
        ///</summary>
        public TestContext TestContext
        {
            get
            {
                return testContextInstance;
            }
            set
            {
                testContextInstance = value;
            }
        }

        #region Additional test attributes
        // 
        //You can use the following additional attributes as you write your tests:
        //
        //Use ClassInitialize to run code before running the first test in the class
        //[ClassInitialize()]
        //public static void MyClassInitialize(TestContext testContext)
        //{
        //}
        //
        //Use ClassCleanup to run code after all tests in a class have run
        //[ClassCleanup()]
        //public static void MyClassCleanup()
        //{
        //}
        //
        //Use TestInitialize to run code before running each test
        //[TestInitialize()]
        //public void MyTestInitialize()
        //{
        //}
        //
        //Use TestCleanup to run code after each test has run
        //[TestCleanup()]
        //public void MyTestCleanup()
        //{
        //}
        //
        #endregion


        /// <summary>
        ///A test for GroupCommand Constructor
        ///</summary>
        [TestMethod()]
        public void GroupCommandConstructorTest()
        {
            List<Column> groupingColumns = null; // TODO: Initialize to an appropriate value
            Dictionary<Column, Aggregation> aggregatedColumns = null; // TODO: Initialize to an appropriate value
            GroupCommand target = new GroupCommand(groupingColumns, aggregatedColumns);
            Assert.Inconclusive("TODO: Implement code to verify target");
        }

        /// <summary>
        ///A test for Execute
        ///</summary>
        [TestMethod()]
        public void ExecuteTest()
        {
            List<Column> groupingColumns = null; // TODO: Initialize to an appropriate value
            Dictionary<Column, Aggregation> aggregatedColumns = null; // TODO: Initialize to an appropriate value
            GroupCommand target = new GroupCommand(groupingColumns, aggregatedColumns); // TODO: Initialize to an appropriate value
            Table table = null; // TODO: Initialize to an appropriate value
            Table expected = null; // TODO: Initialize to an appropriate value
            Table actual;
            actual = target.Execute(table);
            Assert.AreEqual(expected, actual);
            Assert.Inconclusive("Verify the correctness of this test method.");
        }
    }
}
